% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_handling.R
\name{add_pseudocount}
\alias{add_pseudocount}
\title{Add pseudocount}
\usage{
add_pseudocount(matrix, pc = 1)
}
\arguments{
\item{matrix}{PCM or PFM matrix}

\item{pc}{The pseoudocount that should be added (default 1)}
}
\value{
A tibble containing the input matrix with the added pseudo count.
}
\description{
Add a pseudo count to a PCM or a PFM
}
\details{
...
}
