% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_handling.R
\name{calculate_information_content}
\alias{calculate_information_content}
\title{Calculate information content}
\usage{
calculate_information_content(pfm)
}
\arguments{
\item{pfm}{Position frequency matrix}
}
\value{
Vector of information content values (same length as input PFM)
}
\description{
Calculate the information content at each position of a PFM
}
\details{
The information content for a given PFM is calculated as described in
    Schmidt et al. (2021), https://doi.org/10.1016/j.xcrm.2021.100194.
}
