% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_handling.R
\name{create_pwm_df}
\alias{create_pwm_df}
\title{Create PWM tibble}
\usage{
create_pwm_df(df, min_datapoints, background)
}
\arguments{
\item{df}{Input dataframe containing the "peptide" that should be converted into a PWM,
the "allele" and the "len" which denotes the length of the peptide}

\item{min_datapoints}{Minimum number of datapoints for this length allele combination
to create a PWM}

\item{background}{The background that should be used to create the PWM}
}
\value{
Tibble that contains the PWM for each length allele combination
}
\description{
Create a tibble, that contains for every allele, peptide length a PWM
}
\details{
...
}
