% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_handling.R
\name{pcm2pfm}
\alias{pcm2pfm}
\title{PCM to PFM}
\usage{
pcm2pfm(pcm, background = NULL)
}
\arguments{
\item{pcm}{The position count matrix}

\item{background}{Normalize the frequencies by a background frequency}
}
\value{
A tibble containing position frequency matrix (PFM).
}
\description{
Create a position frequency matrix (PFM) from a position count matrix (PCM)
}
\details{
The frequencies of the characters in the PCM for each position are computed
    and returned as PFM tibble. The frequencies can be normalised by a background
    frequency if needed.
}
