% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_handling.R
\name{pfm2pwm}
\alias{pfm2pwm}
\title{PFM to PWM}
\usage{
pfm2pwm(pfm, background)
}
\arguments{
\item{pfm}{The position frequency matrix}

\item{background}{A named list with the character as name and backgroud frequency for that character as value}
}
\value{
A tibble containing position frequency matrix (PFM).
}
\description{
Create a position weight matrix (PWM) from a position frequency matrix (PFM)
}
\details{
For each position and character, the log likelihood is calculated, to
    get the position weight matrix. This matrix can be used to score other sequences.
}
