% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_handling.R
\name{pwm_max}
\alias{pwm_max}
\title{Get max PWM score}
\usage{
pwm_max(pwm)
}
\arguments{
\item{pwm}{Position frequency matrix}
}
\value{
Maximum score that can be reached for given PWM
}
\description{
Calculate the maximum possible score of an PWM
}
\details{
...
}
