% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_handling.R
\name{score_seq}
\alias{score_seq}
\title{Score sequence}
\usage{
score_seq(seq, pwm)
}
\arguments{
\item{seq}{List of character sequences matching the PWM}

\item{pwm}{Position weight matrix tibble}
}
\value{
Vector of scores
}
\description{
Score a sequence with a position weight matrix (PWM)
}
\details{
For each character in the sequence, the score from the PWM is summed
    up to get the final score
}
