% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_handling.R
\name{score_seq_allele}
\alias{score_seq_allele}
\title{Score sequence allele combination}
\usage{
score_seq_allele(seq, allele, pwm, normalise = T)
}
\arguments{
\item{seq}{List of character sequences matching the PWM}

\item{allele}{List of alleles that have the same length as seq}

\item{pwm}{tibble containing columns allele, peptide length, pwm}

\item{normalise}{If TRUE, the score is normalised by the maximum score of the given PWM}
}
\value{
Vector of scores
}
\description{
Score a sequence allele length combination with a position weight matrix (PWM)
}
\details{
For each character in the sequence, the score from the PWM is summed
    up to get the final score
}
