% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_handling.R
\name{score_seq_allele_pos}
\alias{score_seq_allele_pos}
\title{Score sequence allele combination for specific positions}
\usage{
score_seq_allele_pos(seq, mhc, pos, pwm_tibble, normalise = T)
}
\arguments{
\item{seq}{List of character sequences matching the PWM}

\item{pos}{List of positions}

\item{normalise}{If TRUE, the score is normalised by the maximum score of the given PWM}

\item{allele}{List of alleles that have the same length as seq}

\item{pwm}{tibble containing columns allele, peptide length, pwm}
}
\value{
List of PWM scores at given position
}
\description{
Score a sequence allele length combination with a position weight matrix (PWM) for one position in the given sequence
}
\details{
...
}
