% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_handling.R
\name{score_seq_allele_position_wise}
\alias{score_seq_allele_position_wise}
\title{Score sequence allele combination position wise}
\usage{
score_seq_allele_position_wise(aa, pos, allele, len, pwm, normalise = T)
}
\arguments{
\item{allele}{List of alleles that has the same length as seq}

\item{pwm}{tibble containing columns allele, peptide length, pwm}

\item{normalise}{If TRUE, the score is normalised by the maximum score of the given PWM}

\item{seq}{List of character sequences matching the PWM}
}
\value{
Vector of scores
}
\description{
Score a sequence allele length combination with a position weight matrix (PWM)
}
\details{
For each character in the sequence, the PWM score is added as an extra column
}
