% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_handling.R
\name{seqs2pcm}
\alias{seqs2pcm}
\title{Sequences to PCM}
\usage{
seqs2pcm(sequences, alphabet = amino_acids)
}
\arguments{
\item{sequences}{vector of input sequences}

\item{alphabet}{list of characters that can occure in the sequences (default: 20 amino acids)}
}
\value{
A tibble containing the position count matrix (columns: letters; rows: positions)
}
\description{
Create position count matrix (PCM) form list of equal length strings
}
\details{
For every position the different characters are counted and returned
    as a matrix.
}
