% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_handling.R
\name{seqs2pfm}
\alias{seqs2pfm}
\title{Sequences to PFM}
\usage{
seqs2pfm(sequences, background = NULL, alphabet = amino_acids, pc = 1)
}
\arguments{
\item{sequences}{vector of input sequences}

\item{background}{A named list with the character as name and backgroud frequency for that character as value}

\item{alphabet}{list of characters that can occure in the sequences (default: 20 amino acids)}
}
\value{
A tibble containing the position frequency matrix (columns: letters; rows: positions)
}
\description{
Create position frequency matrix (PFM) form list of equal length strings
}
\details{
The frequency can also be normalised by a background frequency.
}
