% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_handling.R
\name{seqs2pwm}
\alias{seqs2pwm}
\title{Sequences to PWM}
\usage{
seqs2pwm(sequences, background, alphabet = amino_acids, pc = 1)
}
\arguments{
\item{sequences}{vector of input sequences}

\item{background}{A named list with the character as name and backgroud frequency for that character as value}

\item{alphabet}{list of characters that can occure in the sequences (default: 20 amino acids)}

\item{pc}{pseudo count (default 1)}
}
\value{
A tibble containing the position weight matrix (columns: letters; rows: positions)
}
\description{
Create position weight matrix (PWM) form list of equal length strings
}
\details{
...
}
